Security: disable SELinux within the containers
We are currently relabeling the directories in the host's FS with each mount-point, messing up the host's security. It's better to just disable SELinux within the container for now, as that is safer than the current situation. See https://danwalsh.livejournal.com/81143.html

For the already relabeled paths due to running contr before, it's recommendable to run `restorecon -rF /path/to/mounted/dir` to restore the original FS labels. Especially important if system paths like ~/.ssh were mounted as volumes.