Fix PS1 prompt in Debian's dash
Use printf to input ASCII control characters (o001, o033, o002) instead of using escape sequences. These control characters are never printed, so they make for a more robust choice because they do not clutter the prompt with symbols in cases where the escape sequences are not supported, like in Debian's dash.

With this change, even the colors now work in dash.