Improve code readability of function make_volume_noexec
Using substring parameter expansion for conditional pattern matching is an anti-pattern, that I probably invented. Although it's nice for quick one-liners, it's slow to execute and hard to read, hard to understand, specially because the condition looks inverted.

Use case statements for pattern matching instead, which is the idiomatic way.