Change terminal title using PS1 variable inside the container
Add extra Xterm Control Sequences to the PS1 variable we set inside the container to also set the title of the terminal. Avoids the title being stuck at the enormous command we execute to start the container.

Also special-case the PS1 variable in busybox to avoid line-length problems in the interactive shell due to busybox erroneously allocating space for some non-printable control characters.