Security: Published ports bind to localhost only by default
contr is meant primarily for interactive use, not for running public servers, hence, change Podman's default to only bind to 127.0.0.1, instead of binding to all IP addresses by default. Also publish ports to the same port number on the host when the host port is not specified on the command line, and when the container port is higher than 1023.