Detect host's unprivileged ports range from sysctl
We automatically bind container ports to the same port number on the host system, but only for port number 1024 or higher. Instead of hard-coding the number 1024, we now detect the start of the range of ports available to unprivileged users from sysctl.

The unprivileged ports range is ignored if the program is run as root.